Split the code into multiple files for easier viewing and troubleshooting
